Don't handle complicated nesting on chat-response
Sometimes chat-response which we might send it over to _ai/ might be complicated nesting with other additional information like texts and other data-structures. It shouldn't be Fractl that should handle such, so, it must be the responsibility of client to only send the proper fractl code to Fractl side of things.

Signed-off-by: Pratik Karki <pratik@fractl.io>