Update to handle nested dependency specifications
The implementation for handling nested dependency specifications has
been updated and enhanced in several areas.
This change makes sure that the structure of dependencies, regardless
if they're nested or not, is processed correctly.
Moreover, this commit includes the necessary transformations for proper
installation and loading of dependencies.
This leads to improved processing and dependency handling in the system.

The new dependency should be a map of following structure:
```clojure
(quote [{:name "WeatherInfo"
        :type :fractl-model
        :version "1.0.0"
        :source {:type :github
                 :org "fractl-dev"
                 :repo "weather_info"
                 :branch "main"}}])
```

Signed-off-by: Pratik Karki <pratik@fractl.io>