Update dependency model name handling (#1309)
* Update dependency model name handling

Check if the dependency is a map in addition to the existing keyword and
vector checks. This increases the flexibility in dealing with different
data types.

* Update to handle nested dependency specifications

The implementation for handling nested dependency specifications has
been updated and enhanced in several areas.
This change makes sure that the structure of dependencies, regardless
if they're nested or not, is processed correctly.
Moreover, this commit includes the necessary transformations for proper
installation and loading of dependencies.
This leads to improved processing and dependency handling in the system.

The new dependency should be a map of following structure:
```clojure
(quote [{:name "WeatherInfo"
        :type :fractl-model
        :version "1.0.0"
        :source {:type :github
                 :org "fractl-dev"
                 :repo "weather_info"
                 :branch "main"}}])
```

* Fix model-name bug

* Refactor dependency management in build scripts

Removed the usage of the :clj-dependencies key from model specs and
simplified dependency handling in build and package scripts.
The refactoring involves using only the :dependencies key in the model
file which now includes all necessary details directly in a standardized
format.

* Update import order and add dependency fetching

A new function 'get-dependencies' has been added in package.clj file.
It fetches project dependencies from the model provided. The
'project-spec' function has been modified to include these dependencies.

* Fix for load-model issue in Design studio

* Revert earlier change, as the issue is fixed now

The problem was continuation wasn't called after :deps tag was reached
on design studio side of things

* Handle multiple dependencies

---------

Signed-off-by: Pratik Karki <pratik@fractl.io>