Add SHA256 checksum calculation to release workflow
This update amends the automated release workflow by adding steps to
calculate the SHA256 checksums for both the JAR and the standalone
Uberjar. These checksums are then included in the release notes for the
update. This will provide users the ability to verify the integrity of
downloaded files for added security.

Signed-off-by: Pratik Karki <pratik@fractl.io>