Refactor checksum calculation and storage in GitHub workflow
The checksum calculation for the jar and uberjar in our GitHub workflow
has been streamlined. This change allows the checksum to be set directly
as environmental variables for immediate and easier access.
Additionally, the creation of a checksum.txt file is now implemented for
storing checksums for future use and reference.

Signed-off-by: Pratik Karki <pratik@fractl.io>